Fix implementation
I may be missing something, but the implementation didn't seem right.
The output size didn't always match that requested, and there was some
odd mixing of the filter and spatial axes. I've now altered it so now
the spatial axes are simply 'block replicated' in what seems like a
more intuitive and correct manner.